Put new component in own file
